Docs: Clarify `yarn start` behavior in README.md
The `README.md` file for the dispensing app is improved to clarify the behavior of the `yarn start` command.

Currently, the `README.md` only shows the `yarn start` command with arguments to connect to a local backend. This can be misleading for new contributors, as it implies that a local backend setup is required to run the application.

The `README.md` is updated to show that:
1.  `yarn start` can be run without any arguments.
2.  By default, `yarn start` will proxy requests to `https://dev3.openmrs.org`.
3.  The command with arguments can be used to connect to a local backend.

This change will make it easier for new contributors to get started with the project.